Changed t_fixture_id to quint32. Changed Doc::m_fixtureArray into a QMap, eliminating the 128 fixture limitation. Added singleton UI destruction code to ~App() to get rid of segfaults.

+++ engine/test/function_test.cpp
@@ -170,7 +170,7 @@ void Function_Test::slotFixtureRemoved()
 
     QCOMPARE(stub->m_slotFixtureRemovedId, Fixture::invalidId());
     doc.deleteFixture(42);
-    QCOMPARE(stub->m_slotFixtureRemovedId, 42);
+    QCOMPARE(stub->m_slotFixtureRemovedId, quint32(42));
 }
 
 void Function_Test::invalidId()
